feat: add cw init command, XDG support, release infrastructure
Add `cw init` command for interactive and scriptable client setup with
input validation (name, branch, path, git repo, purpose enum). Includes
YAML injection prevention via regex-validated branch names, enum-validated
purposes, and single-quoted path output. Handles empty/malformed config
files gracefully.

Add XDG_CONFIG_HOME/XDG_DATA_HOME support for config and state dirs.
Add GitHub Actions release workflow and portable release.sh script.
Pin CI actions to commit SHAs. Update README with installation and
usage docs, add CONFIG_REFERENCE.md.

Review fixes: extract make_git_repo test fixture to conftest, split
compound test assertions, add coverage for edge cases (empty config,
invalid purposes, malformed YAML), strengthen purpose value assertions,
fix test typing (explicit Path constructors, StrEnum .value comparisons).

+++ src/cw/daemon.py
@@ -557,7 +557,7 @@ async def _wait_for_idle_event(
     client: str,
     purpose: str,
     *,
-    timeout: int = _DAEMON_TASK_TIMEOUT_S,
+    timeout: float = _DAEMON_TASK_TIMEOUT_S,
     poll_interval: float = 2.0,
     shutdown_event: asyncio.Event | None = None,
 ) -> dict[str, object] | None:
